Attempt to abstract out the database connection
This didn't work out as well as I'd hoped because it seems like the
assertion library (chai-http) can't see the database unless it's
explicitly connected to within each of the Controller methods. For now
this means that things are quite verbose, and not much of a line saving
on what was there before, but at least the connection string lives
elsewhere and the database connection is it's own class within the
application.